initial commit of ooh323
Brief history:
=============

H323 is the start of VoIP, and during the time more and more people and
company focusing on SIP but still there’s a lot of legacy devices
talking H323 and those still working well.

FreeSWITCH has mod_opal and mod_h323 implemented h323 support, they
respectively depending on the OPAL and h323plus libs. Those two libs are
both forked from openh323 and is C++ based, and also they are depending
on the ptlib and need specific ptlib version respectively and has the
impression of hard to build.

As more and more people interest in video support we were looking the
possibility of adding that on top of mod_opal or mod_h323, as it seems
complicated to figure out how the modules and libs works exactly we
found another lib during the discussion of
http://lists.freeswitch.org/pipermail/freeswitch-users/2013-February/092265.html
and it is written in C and seems simple.

ooh323c:
=======

ooh323c turns out is a great lib despite that it seems been abandoned,
we took the code from http://sourceforge.net/projects/ooh323c/ and
http://www.obj-sys.com/telephony-objective.shtml (link is dead) and
forked to github.com/traviscross/ooh323c based on the latest version
0.9.4 which is the latest one we could find. We made some updates:

- Some bug fixes
- Added video fast update support
- Added H264 support

mod_ooh323:
==========

The mod is implemented as an endpoint like mod_sofia. It using the
ooh323c library for the h323 signalling and use RTP implementation in
the FreeSWITCH core. And because it depends on the core media feature
introduced in FreeSWITCH 1.4 so it should only work with the latest
FreeSWITCH and no backports to 1.2 is planed.

The config file is conf/autoload_configs/ooh323.conf.xml where some
params can be configured.

Some features we implemented:

- Listens and accepts call on 1720 ( or whatever port configured)
- Audio & Video Support (H263 & H264)
- outbound calls (experimental) with dial string ooh323/user@ip:port
- Register to GK (experimental)

Thanks-To: Zhubo Cheng <wavecb@gmail.com>