Fix Issue #11423: NSG Relay Automation Failed due to unable to find span ID while creating ISUP circuit
Present implemenatation of Transparent E1 Over IP is reuslting in NSG Relay automation
test failure as span Id was getting retrieved, depending upon the ccSpanId as present
in freetdm.conf.xml which is incorrect.

The code chages make sure to find transparent cricuit if present in SS7 module
depending upon the new timeslot "transparent" introduced. This is a configurable
parameter that must be present in freetdm.conf.xml