FS-9555 format/mod_aes_wav
For playback and recording to encrypted wav files as streams. This
removes the need to encrypt or decrypt the entire file in one
batch. Allows for more secure(audio is spending less time decrypted)
audio handling, and more performant(less latency, and workload more
evenly spread over the life of the call).